+IWriter interface
This will be the thing that objects use to serialize their stuff out to
file, but technically this abstraction means they don't know (or care)
where it's going.

This interface also grants formatting (indent in and out) functionality,
which something like a TextWriter or Stream wouldn't do. We could derive
a new stream type, but then we wouldn't have interface mockability.